+++ CHANGELOG.md
@@ -1,9 +1,15 @@
 # Release Notes for Payment Express for Craft Commerce
 
-## 1.0.1 - 2019-02-01
+## 1.1.0 - 2019-02-01
+### Added
+- Added Gateway overview widget.
+> {tip} Add the new widget on your dashboard. 
+By default it looks for a gateway with the handle "paymentExpress" but can be changed on the widgets settings.
 
+## 1.0.1 - 2019-02-01
+### Added
 - Added Craft Plugin Store support.
 
 ## 1.0 - 2019-02-01
-
+### Added
 - Initial release.